Move up down jig popups to separate file
